Create PromptForm component to refactor functionality on App
